% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVARDA_compiled.R
\name{AVARDA_compiled}
\alias{AVARDA_compiled}
\title{Primary function to run AVARDA modules together}
\usage{
AVARDA_compiled(
  case_data,
  blast,
  total_prob,
  pairwise,
  dict,
  threshold = 5,
  cores,
  mod_3
)
}
\arguments{
\item{case_data}{A data.frame of PhIPseq data that identifies hits by some theshold, can be binary hits or some value}

\item{blast}{A data.frame listing the blastp, or tblastn, alignments of PhIP-seq peptides to organisms of interact}

\item{total_prob}{A data.frame listing the degree of representation each virus has in the phip-seq library}

\item{pairwise}{A number.}

\item{dict}{A number.}

\item{threshold}{A number.}

\item{cores}{A number.}

\item{mod_3}{A number.}
}
\value{
Output of AVARDA algorithm; a data frame listing the likelihood of each virus causing in infection in a given sample
}
\description{
Primary function to run AVARDA modules together
}
\examples{
\dontrun{do later}
}
