% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\title{Take a matrix and binary it}
\usage{
binary(matrix, threshold)
}
\arguments{
\item{matrix}{A matrix}

\item{threshold}{A number.}
}
\value{
The binaried version of \code{matrix} setting things above \code{threshold} to 1 and below to zero.
}
\description{
Take a matrix and binary it
}
\examples{
binary(matrix(c(100,50,0,100),nrow = 2), 70)
}
