% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.R
\name{bt}
\alias{bt}
\title{Perform a binomial test}
\usage{
bt(a, b, p = 0.5)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{p}{A number}
}
\value{
Returns the binomial test pvalue for an event that occurred \code{a} times out of a total of \code{b} with a null probability assumption of \code{p}
}
\description{
Perform a binomial test
}
\examples{
bt(2,10,.5)
}
